@@ -1,20 +1,20 @@
 package io.smallrye.graphql.tests.client.dynamic;
 
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.runner.RunWith;
+import org.jboss.arquillian.junit5.ArquillianExtension;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClient;
 import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClientBuilder;
 import io.vertx.ext.web.client.WebClientOptions;
 
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class DynamicClientSingleOperationsOverWebsocketTest extends DynamicClientSingleOperationsTestBase {
 
-    @Before
+    @BeforeEach
     public void prepare() {
         client = (VertxDynamicGraphQLClient) new VertxDynamicGraphQLClientBuilder()
                 .url(testingURL.toString() + "graphql")
@@ -23,7 +23,7 @@ public class DynamicClientSingleOperationsOverWebsocketTest extends DynamicClien
                 .build();
     }
 
-    @After
+    @AfterEach
     public void cleanup() {
         client.close();
     }
