@@ -1,6 +1,6 @@
 package io.smallrye.graphql.tests.client.parsing;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.net.URL;
 import java.util.Calendar;
@@ -8,17 +8,17 @@ import java.util.Date;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.client.vertx.typesafe.VertxTypesafeGraphQLClientBuilder;
 
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class TypesafeClientFormatAnnotationsTest {
 
@@ -33,7 +33,7 @@ public class TypesafeClientFormatAnnotationsTest {
 
     protected FormatAnnotationsClientApi client;
 
-    @Before
+    @BeforeEach
     public void prepare() {
         client = new VertxTypesafeGraphQLClientBuilder()
                 .endpoint(testingURL.toString() + "graphql")
