@@ -9,20 +9,20 @@ import java.net.URL;
 import org.eclipse.microprofile.graphql.Name;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.jandex.Index;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.client.vertx.typesafe.VertxTypesafeGraphQLClientBuilder;
 import io.smallrye.graphql.tests.client.typesafe.directives.model.SomeClassClient;
 import io.smallrye.graphql.tests.client.typesafe.directives.model.SomeClassServer;
 
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class TypesafeStaticDirectivesClientModelTest {
 
@@ -41,7 +41,7 @@ public class TypesafeStaticDirectivesClientModelTest {
     protected ClientApi client;
     private boolean onlyOnce = false;
 
-    @Before
+    @BeforeEach
     public void prepare() {
         if (!onlyOnce) {
             Index index = null;
