@@ -6,10 +6,10 @@ import java.io.Closeable;
 import java.io.IOException;
 
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.jandex.Index;
-import org.junit.Before;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.client.typesafe.api.GraphQLClientApi;
 import io.smallrye.graphql.client.vertx.typesafe.VertxTypesafeGraphQLClientBuilder;
@@ -18,12 +18,12 @@ import io.smallrye.graphql.client.websocket.WebsocketSubprotocol;
 /**
  * Test subscriptions with a typesafe client and graphql-ws subprotocol
  */
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class TypesafeClientGraphQLTransportWSSubscriptionWithClientModelTest extends AbstractTypesafeClientSubscriptionTest {
     private boolean onlyOnce = false;
 
-    @Before
+    @BeforeEach
     public void prepare() {
         if (!onlyOnce) {
             Index index = null;
