@@ -11,19 +11,19 @@ import org.eclipse.microprofile.graphql.Query;
 import org.eclipse.microprofile.graphql.Source;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.config.ConfigKey;
 import io.smallrye.graphql.tests.GraphQLAssured;
 
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class ExcludeNullFieldsInResponseTest {
 
@@ -111,12 +111,12 @@ public class ExcludeNullFieldsInResponseTest {
         assertThat(response).contains("{\"data\":{}}");
     }
 
-    @Before
+    @BeforeEach
     public void setExcludeNullFieldProperty() {
         System.setProperty(ConfigKey.EXCLUDE_NULL_FIELDS_IN_RESPONSES, "true");
     }
 
-    @After
+    @AfterEach
     public void clearExcludeNullFieldProperty() {
         System.clearProperty(ConfigKey.EXCLUDE_NULL_FIELDS_IN_RESPONSES);
     }
