@@ -1,7 +1,7 @@
 package io.smallrye.graphql.tests.metrics;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.net.URL;
 import java.time.Duration;
@@ -11,14 +11,14 @@ import java.util.concurrent.TimeUnit;
 import org.eclipse.microprofile.graphql.GraphQLApi;
 import org.eclipse.microprofile.graphql.Query;
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
@@ -36,7 +36,7 @@ import io.smallrye.mutiny.unchecked.Unchecked;
 /**
  * Verify that metrics are produced as expected.
  */
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 public class MicrometerMetricsTestCase {
 
     @Deployment
@@ -74,7 +74,7 @@ public class MicrometerMetricsTestCase {
 
     DynamicGraphQLClient client;
 
-    @Before
+    @BeforeEach
     public void before() {
         client = new VertxDynamicGraphQLClientBuilder()
                 .url(testingURL.toString() + "graphql")
