@@ -1,26 +1,25 @@
 package io.smallrye.graphql.tests.stork;
 
-import static org.junit.Assert.assertEquals;
-
 import java.net.URL;
 
 import org.eclipse.microprofile.graphql.GraphQLApi;
 import org.eclipse.microprofile.graphql.Query;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.smallrye.graphql.client.typesafe.api.GraphQLClientApi;
 import io.smallrye.graphql.client.vertx.typesafe.VertxTypesafeGraphQLClientBuilder;
 import io.smallrye.stork.Stork;
 
-@RunWith(Arquillian.class)
+@ExtendWith(ArquillianExtension.class)
 @RunAsClient
 public class StorkTest {
 
@@ -35,7 +34,7 @@ public class StorkTest {
 
     private DummyClientApi client;
 
-    @Before
+    @BeforeEach
     public void prepare() {
         try {
             System.out.println("CORRECT TESTING URL = " + testingURL);
@@ -55,7 +54,7 @@ public class StorkTest {
     @Test
     public void callThroughStork() {
         try {
-            assertEquals("ok", client.q());
+            Assertions.assertEquals("ok", client.q());
         } catch (Throwable t) {
             t.printStackTrace();
             throw t;
